refactor: 회원 탈퇴 로직을 회원 가입 로직을 수행하는 클래스와 통합
