--- src/docs/asciidoc/user-withdraw.adoc
@@ -1,15 +1,0 @@
-= User 수정 API
-:doctype: book
-:icons: font
-:source-highlighter: highlightjs
-:toc: left
-:toclevels: 2
-
-== 회원 탈퇴
-=== 1. 회원 탈퇴에 성공한다
-HTTP Request
-include::{snippets}/UserWithdrawApi/Withdraw/Success/http-request.adoc[]
-include::{snippets}/UserWithdrawApi/Withdraw/Success/request-headers.adoc[]
-
-HTTP Response
-include::{snippets}/UserWithdrawApi/Withdraw/Success/http-response.adoc[]
